mask_parameter.py
