+++ src/pint/models/parameter/__init__.py
@@ -1,2 +1,3 @@
 from pint.models.parameter.parameter import *
-from pint.models.parameter.pair_parameter import *
+from pint.models.parameter.pair_parameter import pairParameter
+from pint.models.parameter.mask_parameter import maskParameter
