+++ src/pint/models/parameter/parameter.py
@@ -33,7 +33,6 @@ from loguru import logger as log
 
 from pint import pint_units
 from pint.models import priors
-from pint.observatory import get_observatory
 from pint.pulsar_mjd import (
     Time,
     data2longdouble,
@@ -44,7 +43,6 @@ from pint.pulsar_mjd import (
     time_to_longdouble,
     time_to_mjd_string,
 )
-from pint.toa_select import TOASelect
 from pint.utils import split_prefixed_name
 
 
@@ -72,36 +70,6 @@ def _identity_function(x):
     """
 
     return x
-
-
-def _get_observatory_name(o):
-    """Return observatory name only from an telescope code
-
-    Parameters
-    ----------
-    o : str or unicode
-        Input telescope code
-
-    Returns
-    -------
-    str
-    """
-    return get_observatory(str(o)).name
-
-
-def _return_frequency_asquantity(f):
-    """Return frequency as a quantity (MHz assumed)
-
-    Parameters
-    ----------
-    f : float
-
-    Returns
-    -------
-    astropy.units.Quantity
-    """
-
-    return u.Quantity(f, u.MHz, copy=False)
 
 
 class Parameter:
@@ -1536,583 +1504,3 @@ class prefixParameter:
 
         newpfx = prefixParameter(name=new_name, **kws)
         return newpfx
-
-
-class maskParameter(floatParameter):
-    """Parameter that applies to a subset of TOAs.
-
-    A maskParameter applies to a subset of the TOAs, for example JUMP specifies
-    that their arrival times should be adjusted by the value associated with
-    this JUMP. The criterion is based on either one of the standard fields
-    (telescope, frequency, et cetera) or a flag; and the selection can be on an
-    exact match or on a range.
-
-    Upon creation of a maskParameter, an index part will be added, so that the
-    parameters can be distinguished within the
-    :class:`pint.models.timing_model.TimingModel` object. For example::
-
-        >>> p = maskParameter(name='JUMP', index=2, key="-fe", key_value="G430")
-        >>> p.name
-        'JUMP2'
-
-    The selection criterion can be one of the parameters ``mjd``, ``freq``,
-    ``name``, ``tel`` representing the required columns of a ``.tim`` file, or
-    the name of a flag, starting with ``-``. If the selection criterion is
-    based on ``mjd`` or ``freq`` it is expected to be accompanied by a pair of
-    values that define a range; other criteria are expected to be accompanied
-    by a string that is matched exactly.
-
-    Parameters
-    ----------
-    name : str
-        The name of the parameter.
-    index : int, optional
-        The index number for the prefixed parameter.
-    key : str, optional
-        The key words/flag for the selecting TOAs
-    key_value :  list/single value optional
-        The value for key words/flags. Value can take one value as a flag value.
-        or two value as a range.
-        e.g. ``JUMP freq 430.0 1440.0``. or ``JUMP -fe G430``
-    value : float or np.longdouble, optional
-        Toas/phase adjust value
-    long_double : bool, optional
-        Set float type quantity and value in long double
-    units : str, optional
-        Unit for the offset value
-    description : str, optional
-        Description for the parameter
-    uncertainty: float or np.longdouble
-        uncertainty of the parameter.
-    frozen : bool, optional
-        A flag specifying whether "fitters" should adjust the value of this
-        parameter or leave it fixed.
-    continuous : bool, optional
-        Whether derivatives with respect to this parameter make sense.
-    aliases : list, optional
-        List of aliases for parameter name.
-    """
-
-    # TODO: Is mask parameter provide some other type of parameters other then floatParameter?
-
-    def __init__(
-        self,
-        name,
-        index=1,
-        key=None,
-        key_value=[],
-        value=None,
-        long_double=False,
-        units=None,
-        description=None,
-        uncertainty=None,
-        frozen=True,
-        continuous=False,
-        aliases=[],
-    ):
-        self.is_mask = True
-        # {key_name: (keyvalue parse function, keyvalue length)}
-        # Move this to some other places.
-        self.key_identifier = {
-            "mjd": (float, 2),
-            "freq": (_return_frequency_asquantity, 2),
-            "name": (str, 1),
-            "tel": (_get_observatory_name, 1),
-        }
-
-        if not isinstance(key_value, (list, tuple)):
-            key_value = [key_value]
-
-        # Check key and key value
-        key_value_parser = str
-        if key is not None:
-            if key.lower() in self.key_identifier.keys():
-                key_info = self.key_identifier[key.lower()]
-                if len(key_value) != key_info[1]:
-                    errmsg = f"key {key} takes {key_info[1]} element(s)."
-                    raise ValueError(errmsg)
-                key_value_parser = key_info[0]
-            else:
-                if not key.startswith("-"):
-                    raise ValueError(
-                        "A key to a TOA flag requires a leading '-'."
-                        " Legal keywords that don't require a leading '-' "
-                        "are MJD, FREQ, NAME, TEL."
-                    )
-        self.key = key
-        self.key_value = [
-            key_value_parser(k) for k in key_value
-        ]  # retains string format from .par file to ensure correct data type for comparison
-        self.key_value.sort()
-        self.index = index
-        name_param = name + str(index)
-        self.origin_name = name
-        self.prefix = self.origin_name
-        # Make aliases with index.
-        idx_aliases = []
-        for al in aliases:
-            idx_aliases.append(al + str(self.index))
-        self.prefix_aliases = aliases
-        super().__init__(
-            name=name_param,
-            value=value,
-            units=units,
-            description=description,
-            uncertainty=uncertainty,
-            frozen=frozen,
-            continuous=continuous,
-            aliases=idx_aliases + aliases,
-            long_double=long_double,
-        )
-
-        # For the first mask parameter, add name to aliases for the reading
-        # first mask parameter from parfile.
-        if index == 1:
-            self.aliases.append(name)
-        self.is_prefix = True
-        self._parfile_name = self.origin_name
-
-    def __repr__(self):
-        out = self.__class__.__name__ + "(" + self.name
-        if self.key is not None:
-            out += " " + self.key
-        if self.key_value is not None:
-            for kv in self.key_value:
-                out += " " + str(kv)
-        if self.quantity is not None:
-            out += " " + self.str_quantity(self.quantity)
-        else:
-            out += " " + "UNSET"
-            return out
-        if self.uncertainty is not None and isinstance(self.value, numbers.Number):
-            out += " +/- " + str(self.uncertainty.to(self.units))
-        if self.units is not None:
-            out += " (" + str(self.units) + ")"
-        out += ")"
-
-        return out
-
-    @property
-    def repeatable(self):
-        return True
-
-    def name_matches(self, name):
-        if super().name_matches(name):
-            return True
-        elif self.index == 1:
-            name_idx = name + str(self.index)
-            return super().name_matches(name_idx)
-
-    def from_parfile_line(self, line):
-        """Read mask parameter line (e.g. JUMP).
-
-        Returns
-        -------
-        bool
-            Whether the parfile line is meaningful to this class
-
-        Notes
-        -----
-        The accepted format::
-
-            NAME key key_value parameter_value
-            NAME key key_value parameter_value fit_flag
-            NAME key key_value parameter_value fit_flag uncertainty
-            NAME key key_value parameter_value uncertainty
-            NAME key key_value1 key_value2 parameter_value
-            NAME key key_value1 key_value2 parameter_value fit_flag
-            NAME key key_value1 key_value2 parameter_value fit_flag uncertainty
-            NAME key key_value1 key_value2 parameter_value uncertainty
-
-        where NAME is the name for this class as reported by ``self.name_matches``.
-        """
-        k = line.split()
-        if not k:
-            return False
-        # Test that name matches
-        name = k[0]
-        if not self.name_matches(name):
-            return False
-
-        try:
-            self.key = k[1]
-        except IndexError:
-            raise ValueError(
-                "{}: No key found on timfile line {!r}".format(self.name, line)
-            )
-
-        key_value_info = self.key_identifier.get(self.key.lower(), (str, 1))
-        len_key_v = key_value_info[1]
-        if len(k) < 3 + len_key_v:
-            raise ValueError(
-                "{}: Expected at least {} entries on timfile line {!r}".format(
-                    self.name, 3 + len_key_v, line
-                )
-            )
-
-        for ii in range(len_key_v):
-            if key_value_info[0] != str:
-                try:
-                    kval = float(k[2 + ii])
-                except ValueError:
-                    kval = k[2 + ii]
-            else:
-                kval = k[2 + ii]
-            if ii > len(self.key_value) - 1:
-                self.key_value.append(key_value_info[0](kval))
-            else:
-                self.key_value[ii] = key_value_info[0](kval)
-        if len(k) >= 3 + len_key_v:
-            self.value = k[2 + len_key_v]
-        if len(k) >= 4 + len_key_v:
-            try:
-                fit_flag = int(k[3 + len_key_v])
-                if fit_flag == 0:
-                    self.frozen = True
-                    ucty = 0.0
-                elif fit_flag == 1:
-                    self.frozen = False
-                    ucty = 0.0
-                else:
-                    ucty = fit_flag
-            except ValueError:
-                try:
-                    str2longdouble(k[3 + len_key_v])
-                    ucty = k[3 + len_key_v]
-                except ValueError:
-                    errmsg = "Unidentified string " + k[3 + len_key_v] + " in"
-                    errmsg += " parfile line " + k
-                    raise ValueError(errmsg)
-
-            if len(k) >= 5 + len_key_v:
-                ucty = k[4 + len_key_v]
-            self.uncertainty = self._set_uncertainty(ucty)
-        return True
-
-    def as_parfile_line(self, format="pint"):
-        assert (
-            format.lower() in _parfile_formats
-        ), "parfile format must be one of %s" % ", ".join(
-            ['"%s"' % x for x in _parfile_formats]
-        )
-
-        if self.quantity is None:
-            return ""
-        if self.use_alias is None:
-            name = self.origin_name
-        else:
-            name = self.use_alias
-
-        # special cases for parameter names that change depending on format
-        if name == "EFAC" and not (format.lower() == "pint"):
-            # change to T2EFAC for TEMPO/TEMPO2
-            name = "T2EFAC"
-        elif name == "EQUAD" and not (format.lower() == "pint"):
-            # change to T2EQUAD for TEMPO/TEMPO2
-            name = "T2EQUAD"
-
-        line = "%-15s %s " % (name, self.key)
-        for kv in self.key_value:
-            if isinstance(kv, time.Time):
-                line += f"{time_to_mjd_string(kv)} "
-            elif isinstance(kv, u.Quantity):
-                line += f"{kv.value} "
-            else:
-                line += f"{kv} "
-        line += "%25s" % self.str_quantity(self.quantity)
-        if self.uncertainty is not None:
-            line += " %d %s" % (0 if self.frozen else 1, str(self.uncertainty_value))
-        elif not self.frozen:
-            line += " 1"
-        return line + "\n"
-
-    def new_param(self, index, copy_all=False):
-        """Create a new but same style mask parameter"""
-        if not copy_all:
-            new_mask_param = maskParameter(
-                name=self.origin_name,
-                index=index,
-                long_double=self.long_double,
-                units=self.units,
-                aliases=self.prefix_aliases,
-            )
-        else:
-            new_mask_param = maskParameter(
-                name=self.origin_name,
-                index=index,
-                key=self.key,
-                key_value=self.key_value,
-                value=self.value,
-                long_double=self.long_double,
-                units=self.units,
-                description=self.description,
-                uncertainty=self.uncertainty,
-                frozen=self.frozen,
-                continuous=self.continuous,
-                aliases=self.prefix_aliases,
-            )
-        return new_mask_param
-
-    def select_toa_mask(self, toas):
-        """Select the toas that match the mask.
-
-        Parameters
-        ----------
-        toas: :class:`pint.toas.TOAs`
-
-        Returns
-        -------
-        array
-            An array of TOA indices selected by the mask.
-        """
-        column_match = {"mjd": "mjd_float", "freq": "freq", "tel": "obs"}
-        if len(self.key_value) == 1:
-            if not hasattr(self, "toa_selector"):
-                self.toa_selector = TOASelect(is_range=False, use_hash=True)
-            condition = {self.name: self.key_value[0]}
-        elif len(self.key_value) == 2:
-            if not hasattr(self, "toa_selector"):
-                self.toa_selector = TOASelect(is_range=True, use_hash=True)
-            condition = {self.name: tuple(self.key_value)}
-        elif len(self.key_value) == 0:
-            return np.array([], dtype=int)
-        else:
-            raise ValueError(
-                "Parameter %s has more key values than "
-                "expected.(Expect 1 or 2 key values)" % self.name
-            )
-        # get the table columns
-        # TODO Right now it is only supports mjd, freq, tel, and flagkeys,
-        # We need to consider some more complicated situation
-        if self.key.startswith("-"):
-            key = self.key[1::]
-        else:
-            key = self.key
-
-        tbl = toas.table
-        if (
-            self.key.lower() not in column_match
-        ):  # This only works for the one with flags.
-            # The flags are recomputed every time. If don't
-            # recompute, flags can only be added to the toa table once and then never update,
-            # making it impossible to add additional jump parameters after the par file is read in (pintk)
-            flag_col = [x.get(key, None) for x in tbl["flags"]]
-            tbl[key] = flag_col
-            col = tbl[key]
-        else:
-            col = tbl[column_match[key.lower()]]
-        select_idx = self.toa_selector.get_select_index(condition, col)
-        return select_idx[self.name]
-
-    def compare_key_value(self, other_param):
-        """Compare if the key and value are the same with the other parameter.
-
-        Parameters
-        ----------
-        other_param: maskParameter
-            The parameter to compare.
-
-        Returns
-        -------
-        bool:
-            If the key and value are the same, return True, otherwise False.
-
-        Raises
-        ------
-        ValueError:
-            If the parameter to compare does not have 'key' or 'key_value'.
-        """
-        if not (hasattr(other_param, "key") or hasattr(other_param, "key_value")):
-            raise ValueError("Parameter to compare does not have `key` or `key_value`.")
-        if self.key != other_param.key:
-            return False
-        if self.key_value != other_param.key_value:
-            return False
-        return True
-
-
-class pairParameter(floatParameter):
-    """Parameter type for parameters that need two input floats.
-
-    One example are WAVE parameters.
-
-    Parameters
-    ----------
-    name : str
-        The name of the parameter.
-    value : astropy Time, str, float in mjd, str in mjd.
-        The input parameter MJD value.
-    description : str, optional
-        A short description of what this parameter means.
-    uncertainty : number
-        Current uncertainty of the value.
-    frozen : bool, optional
-        A flag specifying whether "fitters" should adjust the value of this
-        parameter or leave it fixed.
-    continuous : bool, optional, default True
-        A flag specifying whether phase derivatives with respect to this
-        parameter exist.
-    aliases : str, optional
-        List of aliases for the current parameter
-
-    """
-
-    def __init__(
-        self,
-        name=None,
-        index=None,
-        value=None,
-        long_double=False,
-        units=None,
-        description=None,
-        uncertainty=None,
-        frozen=True,
-        continuous=False,
-        aliases=[],
-        **kwargs,
-    ):
-        self.index = index
-        name_param = name
-        self.origin_name = name
-        self.prefix = self.origin_name
-
-        self.prefix_aliases = aliases
-
-        super().__init__(
-            name=name_param,
-            value=value,
-            units=units,
-            description=description,
-            uncertainty=uncertainty,
-            frozen=frozen,
-            continuous=continuous,
-            aliases=aliases,
-            long_double=long_double,
-            **kwargs,
-        )
-
-        self.is_prefix = True
-
-    def name_matches(self, name):
-        if super().name_matches(name):
-            return True
-        else:
-            name_idx = name + str(self.index)
-            return super().name_matches(name_idx)
-
-    def from_parfile_line(self, line):
-        """Read mask parameter line (e.g. JUMP).
-
-        Notes
-        -----
-        The accepted format:
-            NAME value_a value_b
-
-        """
-        try:
-            k = line.split()
-            name = k[0].upper()
-        except IndexError:
-            return False
-        # Test that name matches
-        if not self.name_matches(name):
-            return False
-        try:
-            self.value = (k[1], k[2])
-        except IndexError:
-            return False
-        if name != self.name:
-            # FIXME: what about prefix/mask parameters?
-            self.use_alias = name
-
-        return True
-
-    def as_parfile_line(self, format="pint"):
-        quantity = self.quantity
-        if self.quantity is None:
-            return ""
-        if self.use_alias is None:
-            name = self.name
-        else:
-            name = self.use_alias
-        line = "%-15s " % name
-        line += "%25s" % self.str_quantity(quantity[0])
-        line += " %25s" % self.str_quantity(quantity[1])
-
-        return line + "\n"
-
-    def new_param(self, index):
-        """Create a new but same style mask parameter."""
-        new_pair_param = pairParameter(
-            name=self.origin_name,
-            index=index,
-            long_double=self.long_double,
-            units=self.units,
-            aliases=self.prefix_aliases,
-        )
-        return new_pair_param
-
-    def _set_quantity(self, vals):
-        vals = [floatParameter._set_quantity(self, val) for val in vals]
-        return vals
-
-    def _set_uncertainty(self, vals):
-        return self._set_quantity(vals)
-
-    @property
-    def value(self):
-        """Return the pure value of a parameter.
-
-        This value will associate with parameter default value, which is .units attribute.
-        """
-        if self._quantity is None:
-            return None
-        else:
-            return self._get_value(self._quantity)
-
-    @value.setter
-    def value(self, val):
-        """Method to set .value.
-
-        Setting .value attribute will change the .quantity attribute other than .value attribute.
-        """
-        if val is None:
-            if (
-                not isinstance(self.quantity, (str, bool))
-                and self._quantity is not None
-            ):
-                raise ValueError(
-                    "Setting .value to None will lose the parameter value."
-                )
-            else:
-                self.value = val
-        self._quantity = self._set_quantity(val)
-
-    def str_quantity(self, quan):
-        """Return quantity as a string."""
-        try:
-            # Maybe it's a singleton quantity
-            return floatParameter.str_quantity(self, quan)
-        except AttributeError:
-            # Not a quantity, let's hope it's a list of length two?
-            if len(quan) != 2:
-                raise ValueError("Don't know how to print this as a pair: %s" % (quan,))
-
-        v0 = quan[0].to(self.units).value
-        v1 = quan[1].to(self.units).value
-        if self._long_double:
-            if not isinstance(v0, np.longdouble):
-                raise TypeError(
-                    "Parameter {} is supposed to contain long doubles but contains a float".format(
-                        self
-                    )
-                )
-            if not isinstance(v1, np.longdouble):
-                raise TypeError(
-                    "Parameter {} is supposed to contain long doubles but contains a float".format(
-                        self
-                    )
-                )
-        quan0 = str(v0)
-        quan1 = str(v1)
-        return quan0 + " " + quan1
